Create Design based on LED
This is the space command console control system created for the future aerospace 
system. I guess people's communication will not be as complex as it is now which is 
still needs to use a lot of texts and words to describe the intention in the future. 
Especially in the control system, the specific indication function is fixed. Therefore, the 
system I designed is based on different indicator colors to distinguish different control 
functions. As the working positions of the space station are relatively scattered, based 
on different positions, the display system can have lights flashing at different positions 
(which needs to be unified with the distribution of the receiving station and personnel 
of the space station), so I selected six colors, and the user can give the functions and 
definitions of these colors. In addition, there are three main switches. The first switch 
is the main control, This switch can monitor and operate all LED lights. The other two 
are prepared for lower level commanders. When the system is running, the light next 
to the switch is on, which means that the commander has been authorized to use the 
system. By the way, people can also add more switches, colors and grant more 
controllers based on this logic.