Reflection Exercise
Our body signals can be transformed into a data input to the processor 
through various sensors. Through the operation of the processor, our behavior 
will get a feedback, such as action, sound, pressure, motion frequency, etc. It 
depends on what kind of sensor and what achievement you want to achieve. 
So this possibility is huge, so I need to know a lot of sensor performance 
before I can decide what kind of work I can output. Recently, I have been 
studying the related fields of interactive games. One idea is to formulate their 
virtual action image according to the players' interactive actions. I understand 
the performance of tilt sensor. I can install tilt sensor on players to get their 
tilt angle, and give them different light composition through these data to 
complete the interaction. This is a concrete example. However, these tools 
make my creation more possible, such as using temperature sensors to make 
new crown warning machines and using force sensitive resistors to make 
interactive devices about force. I think it will take some research time.
These sensors are usually a combination of simple metals and conductive 
materials. They usually change their shape through environmental changes, so 
as to achieve the purpose of data transmission, which has an impact on all 
users and producers.
For the method of generating interface, for the project of interactive game, I 
have come up with two modes at this stage. One is through the interface of 
mobile phone and the link of sensor. This means that the screen of the mobile 
phone is the total control of the player. Any game related content and 
introduction are still transmitted to the player in a relatively traditional way, but 
sensors are added in the process of game experience. But I think this is still 
not a revolutionary change. The second way I prefer but need more exploration 
is to directly use sensors to complete the game, but the biggest problem in 
this process is how to make players fully understand the intention of the game. 
I think it may require more sensors or more requirements for the design of the 
game itself. This is what I am currently exploring.